GH-696 extracted to file
